requirements.txt
cython==3.0.5
h5py==3.9.0
imgaug==0.4.0
ipython==7.34.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
keras==2.10
matplotlib==3.7.1
numpy==1.23.5
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
pillow==9.4.0
scikit-image==0.19.3
scipy==1.11.3
tensorflow-gpu==2.10
tensorboard==2.10.0
memory_profiler
imgaug
tqdm
noteboook
jupyter
jupyter-core